Tests FTW
Wanted to make sure all the properties could be set correctly and
used so here is a test suite to do so.

So to do so I used gradle test kit to be able to run gradle task and
inject the current version of the plugin code. And run it with all the
different options. Validating as it goes.

Some of the tests were not possible without forkProperties so that
helped.

The strategy for testing is there is a stubbed out spring boot app under
src/test/resources. For each test case it will copy that to a temp dir,
execute the gradle tasks for the plugin passing different argument and
validate the results.