Add ability to generate multiple OpenAPI docs
This commit adds a new extension property `groupedApiMappings` allowing
a user of this plugin to specify multiple mappings of api url to file
name. This is intended for use in projects utilising the springdoc
GroupedOpenApi feature.

When the `groupedApiMappings` property is specified the single api
properties `apiDocsUrl` and `outputFileName` are ignored. The plugin
attempts to download each mapping in turn.

Fixes #40