Make plugin work under Gradle 7
Gradle 7 requires that task properties be annotated to indicate how they should be
treated when calculating what artifacts need to be rebuilt. One of the methods in
the gradle-processes plugin was missing an annotation, which caused the plugin to
bomb out under Gradle 7.

Fix it by subclassing the task in question and annotating the method.